Moved the skillAbbreviation enum to the chatcommand because of its specific use
